Enrich operation failed state.
In particular fuse into it unwritten file failures so that those
become part of the build log. We also have another case for the
error message of operation failures. This will allow to get
rid of most of the result field specific to operations.

This commit only adds the type and propagates the info into
it. Simplifications are for the next commit.