File_cache: forget about hard links, always copy.
Hard links are problematic for iterated writes (fixed points) and
cache safety on non-clean builders: it's easy to change a cache key's
files by writing to a file that has been hard-linked in the build dir.
This brittleness is unwanted.

Copying is slower but measurement on cold and nil rebuilds do not show
egrerious slowdown as far as end-user usability is concerned. There
is still room for improvement in the future: trash resurection and
copy on write clonefile(2).